fix(driver): pointer typedefs, offsetof guard, ULONG cast, disable INF verifier in CI

+++ driver/src/l2cap_stream.c
@@ -28,7 +28,7 @@ NTSTATUS L2capSendSignaling(
     UNREFERENCED_PARAMETER(DevExt);
     UNREFERENCED_PARAMETER(Data);
     UNREFERENCED_PARAMETER(Length);
-    KdPrint(("OpenWinBlue: L2capSendSignaling len=%u (stub)\n", (UINT)Length));
+    KdPrint(("OpenWinBlue: L2capSendSignaling len=%u (stub)\n", (ULONG)Length));
     return STATUS_SUCCESS;
 }
